Update tax post
Added more information about taxes for both China and Thailand.
Took into account of corporate and individual tax.
Showed the calculation steps with result in red for RMB, and blue for
THB for clearer visual at a glance.

In conclusion, tax in China is less in total compared to Thailand for
both cases if doing business in special economic area. But only 5% more
in payable tax for corporate tax in China if not in such area.